Combine login and registration into auth files (AuthController and AuthService); controller updated to use mapper
