+++ backend/src/main/java/com/codify/backend/service/AuthResult.java
@@ -1,0 +1,8 @@
+package com.codify.backend.service;
+
+import com.codify.backend.model.User;
+
+public record AuthResult(
+	User user,
+	String token
+){}
